requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Image processing
Pillow>=10.4.0
imagehash>=4.3.1
opencv-python-headless>=4.9.0.80
numpy>=1.26.3

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1
